+++ bin/bin.go
@@ -4,6 +4,11 @@
 // Package bin is sample command application lib, provides functional and classic style Apis.
 package bin
 
+import (
+	"regexp"
+	"strings"
+)
+
 const (
 	AppMethodInit     = "Init"
 	AppMethodRun      = "Run"
@@ -19,3 +24,21 @@ const (
 	CmdApp initIota = iota
 	CmdFunc
 )
+
+// Cmd2StringMap command string turn to map string, for standard go method name.
+// like:
+// 		`get-videos` -> `GetVideos`
+// 		`get_videos` -> `GetVideos`
+func Cmd2StringMap(c string) string {
+	reg := regexp.MustCompile(`([-_]+)|(\s{2,})`)
+	c = reg.ReplaceAllString(c, " ")
+
+	var words []string
+	for _, v := range strings.Split(c, " ") {
+		if v != "" {
+			words = append(words, strings.Title(v))
+		}
+	}
+
+	return strings.Join(words, "")
+}